\name{generate_matrix}
\alias{generate_matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
generate matrix files
}
\description{
generate matrix files from HiC-Pro result
can be used as following command:
generate_matrix(all_hic_file="demo.matrix",all_bed_file="demo.bed",outputpdf="FALSE",matrix_dir="hg19",chrom_file="chrom_hg19.sizes")
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{all_hic_file}{
the path of hic matrix file generated by HiC-Pro, can be found in the ../hic_result/matrix/myhic/iced/resolution/**_iced.matrix
}
  \item{all_bed_file}{
the path of hic index file generated by HiC-Pro, can be found in the ../hic_result/matrix/myhic/raw/resolution/**_abs.bed
}
  \item{outputpdf}{
the picture format of result, if set FALSE, result will be plotted at JPEG format, if TRUE, result will be ploted at PDF format
}
  \item{matrix_dir}{
the output path of this function.
}
  \item{chrom_file}{
the chromatin information file
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
