% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ID3}
\alias{ID3}
\title{ID3 - an R/Weka Classifier Trees}
\format{\preformatted{function (formula, data, subset, na.action, control = Weka_control(), options = NULL)  
 - attr(*, "class")= chr [1:2] "R_Weka_classifier_interface" "R_Weka_interface"
 - attr(*, "meta")=List of 4
  ..$ name : chr "weka/classifiers/trees/Id3"
  ..$ kind : chr "R_Weka_classifier_interface"
  ..$ class: chr "Weka_classifier"
  ..$ init : NULL
}}
\source{
Heavily inspired by the code in the
function \link[RWeka]{J48}, and the help of Ista Zahn.
}
\usage{
ID3
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.}

\item{data}{an optional data frame containing the variables in the model.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. See \link{model.frame} for details.}

\item{control}{an object of class \link{Weka_control} giving options to be passed to the Weka learner. Available options can be obtained on-line using the Weka Option Wizard \link{WOW}, or the Weka documentation.}

\item{...}{not used.}
}
\value{
A list inheriting from classes Weka_tree and Weka_classifiers.
See \link[RWeka]{J48} for a list of the components.
}
\description{
ID3 - an R/Weka Classifier Trees
}
\examples{
\dontrun{
 library(RWeka)
 DF2 <- read.arff(system.file("arff", "contact-lenses.arff", package = "RWeka"))
 load_simpleEducationalLearningSchemes()
 ID3(`contact-lenses` ~ ., data = DF2)
}
}
\seealso{
\link[RWeka]{J48}
}
\keyword{datasets}

