\name{if_distribution_analysis}
\alias{if_distribution_analysis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
analyse the interaction frequency distribution
}
\description{
analyse the interaction frequency distribution
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{all_hic_file}{
the path of hic matrix file generated by HiC-Pro, can be found in the ../hic_result/matrix/myhic/iced/resolution/**_iced.matrix
}
  \item{all_bed_file}{
the path of hic index file generated by HiC-Pro, can be found in the ../hic_result/matrix/myhic/raw/resolution/**_abs.bed
}
  \item{bedFile}{
the path of specific sites, should use BED format
}
  \item{matrix_dir}{
the path of matrix, which is output by function generate_matirx()
}
  \item{inter_chromfile}{
the inter-chromatin interaction file generated by generate_matrix(), when set NULL, it will be set as matrix_dir/inter_chrom.iam.
}
  \item{groupNum}{
the random group number. Default is 50
}
  \item{random_analysis}{
Whether run random analysis, Default is TRUE
}
  \item{threshold_percent}{
the up-threshold of Hi-C interactions, part of the largest interaction number will be deleted. Default is 0.005
}
  \item{if_bin_number}{
interacion frequency bin number. all interacions will be sort to these bins according to their interacion number. Default is 20
}
  \item{outputpdf}{
the picture format of result, if set FALSE, result will be plotted at JPEG format, if TRUE, result will be ploted at PDF format
}
  \item{resolution}{
the resolution of Hi-C dataset.
}
  \item{chrom_file}{
the chromatin information file
}
  \item{slide_window}{
if set TRUE, there will use a slide window to make picture smooth
}
}
\details{
HBP sets up and down thresholds, and discards the interaction-pairs whose frequencies are lower than the down threshold or higher than the up threshold. Then HBP separates chromatin interaction frequencies into many bins (referred to as frequency bin), and computes density for specific sites in each frequency bins. Meanwhile, HBP will compare the results with that of random control groups which were generated according to the distribution of these imported specific sites. Based on these imported specific region, every member of random control groups was calculated and made in the similar region. This comparison results make it easier for users to ascertain whether the observed interactions were biologically significant and whether these specific sites have effect on the conformation of chromatin or not.

can be used as following command:
if_distribution_analysis(all_hic_file="demo.matrix",all_bed_file="demo.bed",bedFile="dm3_mars.bed",matrix_dir="L1_GM12878",resolution=50,chrom_file="chrom_dm3.sizes")


if_distribution_analysis(all_hic_file="demo.matrix",all_bed_file="demo.bed",bedFile="dm3_mars.bed",inter_chromfile=NULL,groupNum=50,random_analysis=TRUE,threshold_percent=0.001,if_bin_number=30,outputpdf=TRUE,matrix_dir="SINE",resolution=100,chrom_file="chrom_hg19.sizes",slide_window=TRUE)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
